elearning/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
scikit-learn>=0.24.0
numpy>=1.19.0
tqdm>=4.62.0
joblib>=1.0.0 
# requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9.3
django-crispy-forms>=1.14.0
Pillow>=8.0.0
whitenoise>=5.3.0
django-debug-toolbar
watchdog
requests
python-dotenv


